Remove msapplication-config and manifest entries from base.html
Removes the final two references to the forums site. The two were removed instead of
copied locally because the manifest pointed to a file which could not be found, and the
msapplication-config could not be retrieved correctly (as both are not present in the forums
and the only reference is the Internet Archive, which fails to show me the file correctly due
to an XML parsing error).