Hide Animation API
Since the animation API is pretty much dead but still lives in the Forge codebase, the solution would be to hide the pages and then remove when Forge 'officially' removes it from their framework. If somebody in-between then wants to revamp the docs into a workable state, all the better.